175: Refactors court report DB tables
Changes court reports tables to new structure, also refactors seed data files
for court-reports
