+++ seed/data/dev/01600_court_reports_calculations.js
@@ -1,4 +1,4 @@
-var tableName = 'court_reports_workload_points_calculation'
+var tableName = 'court_reports_calculations'
 
 exports.seed = function (knex, Promise) {
   var existingCrReportIds
@@ -10,13 +10,13 @@ exports.seed = function (knex, Promise) {
     })
     .then(function (reportIds) {
       existingCrReportIds = reportIds
-      return knex('court_reports_workload_points').select('id').first()
+      return knex('workload_points').select('id').first()
     })
     .then(function (crWorkloadPointsId) {
       currentCrPointsId = crWorkloadPointsId
       return knex('workload_owner')
-        .join('court_reports_workload', 'workload_owner.id', 'court_reports_workload.workload_owner_id')
-        .max('court_reports_workload.id AS id')
+        .join('court_reports', 'workload_owner.id', 'court_reports.workload_owner_id')
+        .max('court_reports.id AS id')
         .groupBy('workload_owner.id')
     })
     .then(function (crWorkloadIds) {
@@ -32,9 +32,9 @@ exports.seed = function (knex, Promise) {
           var crWorkloadId = crWorkloadIds[w]
 
           crWorkloadPointsCalculationsToInsert.push({
-            court_reports_workload_id: crWorkloadId.id,
+            court_reports_id: crWorkloadId.id,
             workload_report_id: crReportId.id,
-            court_reports_workload_points_id: currentCrPointsId.id,
+            workload_points_id: currentCrPointsId.id,
             contracted_hours: 37.5,
             reduction_hours: Math.floor(Math.random() * 6) + 1
           })
