--- seed/data/dev/01600_court_reports_workload_points.js
@@ -1,12 +1,0 @@
-var tableName = 'court_reports_workload_points'
-
-exports.seed = function (knex, Promise) {
-  // Deletes ALL existing entries
-  return knex(tableName).del()
-    .then(function () {
-      // Inserts seed entries
-      return knex(tableName).insert({
-        default_contracted_hours: 37
-      })
-    })
-}
